Added navigation graph.
