+++ app/src/main/res/layout/fragment_laws.xml
@@ -12,7 +12,6 @@
         android:id="@+id/labelLaws"
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
-        android:layout_margin="@dimen/default_layout_margin"
         android:padding="@dimen/default_padding"
         android:text="@string/laws_related_to_women_safety"
         android:textStyle="bold"
@@ -26,7 +25,6 @@
         android:id="@+id/recyclerView"
         android:layout_width="@dimen/recycler_view_width_height"
         android:layout_height="@dimen/recycler_view_width_height"
-        android:layout_margin="@dimen/default_layout_margin"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintTop_toBottomOf="@+id/labelLaws"
         app:layout_constraintStart_toStartOf="parent"
