+++ app/src/main/res/layout/fragment_safety_tips.xml
@@ -12,9 +12,9 @@
         android:id="@+id/labelSafetyTips"
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
-        android:layout_margin="@dimen/default_layout_margin"
         android:padding="@dimen/default_padding"
         android:text="@string/tips_for_women_safety"
+        android:fontFamily="@font/aclonica"
         android:textStyle="bold"
         android:textColor="@color/white"
         android:textSize="@dimen/fragment_label_text_size"
@@ -26,7 +26,6 @@
         android:id="@+id/recyclerView"
         android:layout_width="@dimen/recycler_view_width_height"
         android:layout_height="@dimen/recycler_view_width_height"
-        android:layout_margin="@dimen/default_layout_margin"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintTop_toBottomOf="@id/labelSafetyTips"
         app:layout_constraintStart_toStartOf="parent"
