+++ app/src/main/res/layout/law_item.xml
@@ -7,7 +7,6 @@
     <androidx.cardview.widget.CardView
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
-        android:layout_margin="@dimen/default_layout_margin"
         app:cardUseCompatPadding="true"
         app:layout_constraintLeft_toLeftOf="parent"
         app:layout_constraintRight_toRightOf="parent"
@@ -25,6 +24,7 @@
                 android:layout_margin="@dimen/default_layout_margin"
                 android:textSize="@dimen/fragment_text_size"
                 android:textStyle="bold"
+                android:layout_marginStart="@dimen/default_layout_margin"
                 app:layout_constraintLeft_toLeftOf="parent"
                 app:layout_constraintRight_toRightOf="parent"
                 app:layout_constraintTop_toTopOf="parent" />
@@ -33,8 +33,8 @@
                 android:id="@+id/btnExpandLawDescription"
                 android:layout_width="@dimen/image_button_width_height"
                 android:layout_height="@dimen/image_button_width_height"
-                android:layout_marginLeft="10dp"
-                android:layout_marginRight="10dp"
+                android:layout_marginLeft="@dimen/default_layout_margin"
+                android:layout_marginRight="@dimen/default_layout_margin"
                 android:src="@drawable/ic_arrow_down"
                 app:layout_constraintBottom_toBottomOf="@+id/lawTitle"
                 app:layout_constraintEnd_toEndOf="@+id/lawTitle"
