+++ app/src/main/res/values/strings.xml
@@ -75,4 +75,5 @@
     <string name="national_commission_act_description">The National Commission for Women (NCW) is a statutory body of the Government of India, established in January 1992. Lalitha Kumaramangalam was appointed its Chairperson in 2014. The NCW represents the rights of women in India and provides a voice for their issues and concerns. The National Commission for Women Act aims to improve the status of women and worked for their economic empowerment.</string>
     <string name="equal_remuneration_act">Equal Remuneration Act, 1976 </string>
     <string name="equal_remuneration_act_description">This Act prevents discrimination in terms of remuneration. It provides for payment of equal recompense to men and women workers.</string>
+    <string name="hi">Hi</string>
 </resources>
